fix: cast tensors to float32 before numpy in logging helpers
bf16-mixed precision produces BFloat16 tensors, which numpy can't
convert (TypeError: unsupported ScalarType BFloat16). Cast with
.float() before .cpu().numpy() in log_image/log_sequence/log_video so
image/video logging works under mixed precision.

+++ rl4seg3d/utils/logging_helper.py
@@ -44,7 +44,8 @@ def put_text_to_image(img, text):
 
 
 def log_image(logger, img, title, number=0, img_text=None, epoch=0):
-    img = img.cpu().numpy()
+    # .float() so bf16/fp16 tensors (autocast/mixed precision) survive the numpy cast
+    img = img.float().cpu().numpy()
     if logger is None:
         return
     img = (img.copy() * (255 / max(img.max(), 1))).astype(np.uint8).squeeze(0)
@@ -63,7 +64,8 @@ def log_image(logger, img, title, number=0, img_text=None, epoch=0):
 
 
 def log_sequence(logger, img, title, number=0, img_text=None, epoch=0):
-    img = img.cpu().numpy()
+    # .float() so bf16/fp16 tensors (autocast/mixed precision) survive the numpy cast
+    img = img.float().cpu().numpy()
 
     # concat images together into bigger image, maximum 4
     img = np.concatenate([img[..., i].transpose((0, 2, 1)) for i in range(min(img.shape[-1], 4))], axis=2)
@@ -94,7 +96,8 @@ def log_video(logger, img, title, background=None, number=0, img_text=None, epoc
     if logger is None:
         return
     if not isinstance(img, np.ndarray):
-        img = img.cpu().numpy()
+        # .float() so bf16/fp16 tensors (autocast/mixed precision) survive the numpy cast
+        img = img.float().cpu().numpy()
     img = img.transpose((0, 2, 1, 3))
     img = (img.copy() * (255 / max(img.max(), 1))).astype(np.uint8)
     if img_text:
@@ -102,7 +105,8 @@ def log_video(logger, img, title, background=None, number=0, img_text=None, epoc
 
     if background is not None:
         if not isinstance(background, np.ndarray):
-            background = background.cpu().numpy()
+            # .float() so bf16/fp16 tensors (autocast/mixed precision) survive the numpy cast
+            background = background.float().cpu().numpy()
         background = background.transpose((0, 2, 1, 3))
         background = (background.copy() * (255 / max(background.max(), 1))).astype(np.uint8)
         if img_text:
